# torrent/src/http/openapi_generated_api.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package http

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Drop all torrents
	// (DELETE /torrents)
	DropAllTorrents(w http.ResponseWriter, r *http.Request, params DropAllTorrentsParams)
	// List torrents
	// (GET /torrents)
	ListTorrents(w http.ResponseWriter, r *http.Request)
	// Add torrent
	// (POST /torrents)
	AddTorrent(w http.ResponseWriter, r *http.Request)
	// Get torrent
	// (GET /torrents/{infoHash})
	GetTorrent(w http.ResponseWriter, r *http.Request, infoHash InfoHash)
	// Get torrent file stats
	// (GET /torrents/{infoHash}/files/{fileIndex}/stats)
	GetTorrentFileStats(w http.ResponseWriter, r *http.Request, infoHash InfoHash, fileIndex FileIndex)
	// Stream file
	// (GET /torrents/{infoHash}/files/{fileIndex}/stream/{fileName})
	StreamFile(w http.ResponseWriter, r *http.Request, infoHash InfoHash, fileIndex FileIndex, fileName FileName)
	// Get torrent stats
	// (GET /torrents/{infoHash}/stats)
	GetTorrentStats(w http.ResponseWriter, r *http.Request, infoHash InfoHash)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Drop all torrents
// (DELETE /torrents)
func (_ Unimplemented) DropAllTorrents(w http.ResponseWriter, r *http.Request, params DropAllTorrentsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List torrents
// (GET /torrents)
func (_ Unimplemented) ListTorrents(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Add torrent
// (POST /torrents)
func (_ Unimplemented) AddTorrent(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get torrent
// (GET /torrents/{infoHash})
func (_ Unimplemented) GetTorrent(w http.ResponseWriter, r *http.Request, infoHash InfoHash) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get torrent file stats
// (GET /torrents/{infoHash}/files/{fileIndex}/stats)
func (_ Unimplemented) GetTorrentFileStats(w http.ResponseWriter, r *http.Request, infoHash InfoHash, fileIndex FileIndex) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Stream file
// (GET /torrents/{infoHash}/files/{fileIndex}/stream/{fileName})
func (_ Unimplemented) StreamFile(w http.ResponseWriter, r *http.Request, infoHash InfoHash, fileIndex FileIndex, fileName FileName) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get torrent stats
// (GET /torrents/{infoHash}/stats)
func (_ Unimplemented) GetTorrentStats(w http.ResponseWriter, r *http.Request, infoHash InfoHash) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// DropAllTorrents operation middleware
func (siw *ServerInterfaceWrapper) DropAllTorrents(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params DropAllTorrentsParams

	// ------------- Optional query parameter "delete" -------------

	err = runtime.BindQueryParameter("form", true, false, "delete", r.URL.Query(), &params.Delete)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "delete", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DropAllTorrents(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListTorrents operation middleware
func (siw *ServerInterfaceWrapper) ListTorrents(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListTorrents(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AddTorrent operation middleware
func (siw *ServerInterfaceWrapper) AddTorrent(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AddTorrent(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTorrent operation middleware
func (siw *ServerInterfaceWrapper) GetTorrent(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "infoHash" -------------
	var infoHash InfoHash

	err = runtime.BindStyledParameterWithOptions("simple", "infoHash", chi.URLParam(r, "infoHash"), &infoHash, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "infoHash", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTorrent(w, r, infoHash)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTorrentFileStats operation middleware
func (siw *ServerInterfaceWrapper) GetTorrentFileStats(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "infoHash" -------------
	var infoHash InfoHash

	err = runtime.BindStyledParameterWithOptions("simple", "infoHash", chi.URLParam(r, "infoHash"), &infoHash, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "infoHash", Err: err})
		return
	}

	// ------------- Path parameter "fileIndex" -------------
	var fileIndex FileIndex

	err = runtime.BindStyledParameterWithOptions("simple", "fileIndex", chi.URLParam(r, "fileIndex"), &fileIndex, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "fileIndex", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTorrentFileStats(w, r, infoHash, fileIndex)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// StreamFile operation middleware
func (siw *ServerInterfaceWrapper) StreamFile(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "infoHash" -------------
	var infoHash InfoHash

	err = runtime.BindStyledParameterWithOptions("simple", "infoHash", chi.URLParam(r, "infoHash"), &infoHash, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "infoHash", Err: err})
		return
	}

	// ------------- Path parameter "fileIndex" -------------
	var fileIndex FileIndex

	err = runtime.BindStyledParameterWithOptions("simple", "fileIndex", chi.URLParam(r, "fileIndex"), &fileIndex, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "fileIndex", Err: err})
		return
	}

	// ------------- Path parameter "fileName" -------------
	var fileName FileName

	err = runtime.BindStyledParameterWithOptions("simple", "fileName", chi.URLParam(r, "fileName"), &fileName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "fileName", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.StreamFile(w, r, infoHash, fileIndex, fileName)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTorrentStats operation middleware
func (siw *ServerInterfaceWrapper) GetTorrentStats(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "infoHash" -------------
	var infoHash InfoHash

	err = runtime.BindStyledParameterWithOptions("simple", "infoHash", chi.URLParam(r, "infoHash"), &infoHash, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "infoHash", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTorrentStats(w, r, infoHash)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/torrents", wrapper.DropAllTorrents)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/torrents", wrapper.ListTorrents)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/torrents", wrapper.AddTorrent)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/torrents/{infoHash}", wrapper.GetTorrent)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/torrents/{infoHash}/files/{fileIndex}/stats", wrapper.GetTorrentFileStats)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/torrents/{infoHash}/files/{fileIndex}/stream/{fileName}", wrapper.StreamFile)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/torrents/{infoHash}/stats", wrapper.GetTorrentStats)
	})

	return r
}